refactor: classify summary dispatch, query
